films_service/requirements.txt
redis==5.0.4
elasticsearch[async]==8.13.2
fastapi==0.111.0
pydantic==2.8.0
pydantic-settings==2.3.4
gunicorn==22.0.0
PyJWT==2.8.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.47b0